Move serde/serialize impls to separate files
